fedimint_client_module

Module module

Module client interface definitions

Modules

init

recovery

Structs

ClientContext

A Client context for a ClientModule M

ClientContextSelfRef

A reference back to itself that the module cacn get from the ClientContext

DynClientModule

FinalClientIface

A final, fully initialized client

IdxRange

A contiguous range of input/output indexes

OutPointRange

Represents a range of output indices for a single transaction

OutPointRangeIter

PrimaryModulePriority

Priority module priority (lower number is higher priority)

Enums

PrimaryModuleSupport

Which amount units this module supports being primary for

Traits

ClientContextIface

A fedimint-client interface exposed to client modules

ClientModule

Fedimint module client

IClientModule

Type-erased version of ClientModule

Type Aliases

ClientModuleRegistry

StateGenerator