fedimint_client_module

Module oplog

Operation log subsystem of the client

Structs

JsonStringed

Json value using string representation as db encoding.

OperationLogEntry

Represents an operation triggered by a user, typically related to sending or receiving money.

OperationOutcome

Represents the outcome of an operation, combining both the outcome value and its timestamp

Enums

UpdateStreamOrOutcome

Either a stream of operation updates if the operation hasn’t finished yet or its outcome otherwise.

Traits

IOperationLog