fedimint_client_module

Module transaction

Structs and interfaces to construct Fedimint transactions

Modules

builder 🔒

sm 🔒

State machine for submitting transactions

Structs

ClientInput

ClientInputBundle

A group of inputs and state machines responsible for driving their state

ClientInputSM

ClientOutput

ClientOutputBundle

ClientOutputSM

TransactionBuilder

TxSubmissionContext

TxSubmissionStatesSM

State machine to (re-)submit a transaction until it is either accepted or rejected by the federation

Enums

NeverClientContext

A fake sm::Context for NeverClientStateMachine

NeverClientStateMachine

A fake sm::State that can actually never happen.

TxSubmissionStates

Constants

TRANSACTION_SUBMISSION_MODULE_INSTANCE

Reserved module instance id used for client-internal state machines

Functions

tx_submission_sm_decoder