fedimint_client_module::transaction

Module builder

Structs

ClientInput

ClientInputBundle

A group of inputs and state machines responsible for driving their state

ClientInputSM

ClientOutput

ClientOutputBundle

ClientOutputSM

TransactionBuilder

Enums

NeverClientContext

A fake sm::Context for NeverClientStateMachine

NeverClientStateMachine

A fake sm::State that can actually never happen.

Functions

find_range_of_matching_items 🔒

Find the range of indexes in an monotonically increasing arr, that is equal to item

state_gen_to_dyn 🔒